backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Cache & Task Queue
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Report Generation
openpyxl==3.1.5
WeasyPrint==63.1

# File Storage
django-storages==1.14.4

# Utilities
python-decouple==3.8
python-dateutil==2.9.0
django-import-export==4.3.3

# Development & Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# Code Quality
ruff==0.8.4
black==24.10.0
isort==5.13.2

# Security
pip-audit==2.7.3
safety==3.2.14
